Refactor: Modularize visualization functions
